+++ test/pub_test.dart
@@ -31,6 +31,10 @@ void main() {
     // Flutter pub get should run in the example project anyways
     // Can't test this with fvm since the output is the same as flutter
     // expectLine(stdout, ['example', 'fvm flutter pub get']);
+    expectLine(
+      stdout,
+      [p.join('fvm_puby_test', 'nested'), 'fvm flutter pub get'],
+    );
 
     // invalid_pubspec
     expectLine(stdout, ['invalid_pubspec_test', 'Error parsing pubspec']);